Adding password protection to celery::redis
By default now redis server will be launched with a randomly generated password
that will be stored in user home directory in a text file only user has
permissions to read. Celery workers load password from that file, and so need to
have access to the home directory (ok on NCI) but also need to be launched after
the file was created.

Password is generated once and then re-used unless user manually deletes the
file where it is stored, `.datacube-redis` at the moment.