Support for multi-band sources
Adding extra field `band` to dataset description document. When set read that
specific band only. Previously `layer` was used as band when it was set to an
integer, however with MODIS data we need both `layer` as a string pointing to
the sub-dataset within a hdf file and also a band within the sub-dataset.