Forcing installation of cython to support netcdf4 1.2.8
cython is a compulsory dependency of netcdf4 starting from 1.2.8, but since it's
a build-time dependency conda doesn't install it, netdcf4 is compiled
externally. Lack of cython doesn't break netcdf4 library, but it does break
compliance checker, because it uses `pkg_resource` lib to check if dependencies
for plugins are installed and `pkg_resource` reports that cython is needed to
import `netcdf4` which isn't true.