Remove ingest's custom product validation
It was doing its own validation and disabling the index's, but its own
had several bugs (such as original/new docs being the wrong way around,
and allowing any metadata type changes).

It was also potentially dangerous to have the validation duplicated in
two places, as a new invariant added in one may not be added in the
other.

(cherry picked from commit 4dd84c6)