Test DriverManager serialization with logging
`DriverManager` tries to serialize the given `Index`. When
the `Index` has a thread lock somewhere inside it, the program
crashes. One way to keep the lock alive is to turn logging on.
Although, nobody knows how that works yet. Need to find a simpler
way to ensure the existence of that lock.