Raise a more general error on being unable to match a query to a product
The way the code is structured makes it somewhat difficult to find the case
that a `product` has been specified explicitly, but that some of the other
search terms are not available on that product.

The magic happens in the _products.py:search_robust() method, which is used
in a few places and I don't want to break the assumptions of other callers.