Remove product matching options from dataset update cli command
Dataset update operates on existing datasets only, so no need to auto-match
anything, every dataset should be already in the database and associated with
some product.

Also dataset update won't load and resolve lineage sources anymore, internal
update apis don't really support adding new lineage data as part of an `update`
call anyway, although it is part of `can_update` check.

What this means is that there is a possibly change in behaviour when new version
of dataset has changes in the lineage document. Previously this would be
reported as incompatible change, now it won't be detected at all.