Remove sources_policy option from `dataset add` #451
- Verification is not part of dataset add anymore
- The default behaviour is what `sources_policy=ensure` used to do
  - add lineage datasets if they are missing,
    but don't check for consistency with the DB for those that are present
- `sources_policy=skip` is achieved with `--no-auto-add-lineage` now

We might re-add verification of lineage option in the future, but it would have
to allow for greater flexibility than previous implementation, at the very least
it should support the same mechanism currently used by "safe to update", except
it will be "divergence from DB for these fields is ok", but this probably needs
to be extended too. What is an ok divergence for product "A" inside the lineage,
might not be ok for product "B".