Simplify eo3 detection: use only $schema
We know that all eo3 documents have a schema if they pass the validator
tool, and all older datasets have no schema. So instead of heuristics,
we can just use the `$schema` field to detect if it's eo3.

It also now rejects unknown schemas. Several of us have indexed eo3
documents accidentally recently, which ODC-1.7 quietly accepts despite
not supporting eo3. It's friendlier to reject them immediately, and
leaves room to match a Stac Item or Eo4 schema in the future.