Add compiled architecture diagrams to repo
Removes the plantuml requirement for building the docs, which in turn
requires Java, which is a very heavyweight extra dependency to have.

This also adds a new 'make' target: 'plantuml_diags', for rebuilding
the diagrams if and when they change.

Getting this makefile working was a bit tricky, and required working
around the 'catch-all' target which Sphinx uses. Make will always search
for a rule to build a file, even if the file already exists, and so was
trying to rebuild the plantuml files using the Sphinx target (FAIL). The
solution is simple, and my knowledge of makefiles is now enhanced.