Clear error messages for archive, restore
A user asked for help on Slack because `datacube dataset archive` was
throwing an `AttributeError`. This turned out to be because one of the
UUIDs provided on the command-line was unknown by datacube.

This adds a friendlier error message for both `archive` and the
`restore` commands.

There's also some clean-up of types to fix some warnings in the affected
methods.

(Note that the error message is output via `stdout` to be consistent
with the rest of the commands.)