Document resampling strategy and overviews (#1017)
Document resampling strategy and overviews

Apparently if stored overviews exist then they will be used, and therefore any subsequent choice of resampling strategy may in effect be silently overridden when down-sampling. 

Consequently, for some applications (depending on the individual data file), it may be inappropriate to allow the API to perform the resampling.