DB hostname comment for Ubuntu based installation
If the DB hostname in the datacube_integration.conf file is set to localhost, the integration tests fail with the message that implies database connectivity issue. This is because for Ubuntu, postgres uses /var/run/postgresql folder. 
The fix was mentioned in this issue: https://github.com/opendatacube/datacube-core/issues/1329