Fix OrderedDict type issues
Modern MyPy gives an incomprehensible
error message for these two
OrderedDicts. Pyright gives
a more readable error:

Argument expression after ** must be a mapping with a "str" key type

so convert the keys of
the dictionary from xr_coords
to strings.

The list of tuples with
the type correct keys cannot
come after the unpacked dictionary
argument to OrderedDict, so
add them through the update method
instead.

This changes the behavior of the code
since there will no longer be
a TypeError exception when the keys
overlap between the two dictionaries.