CI: retry more in setup-miniconda
The action seems to fail intermittently
with:

CondaHTTPError: HTTP 000 CONNECTION FAILED for url <https://conda.anaconda.org/conda-forge/linux-64/repodata.json>
Elapsed: -

An HTTP error occurred when trying to retrieve this URL.
HTTP errors are often intermittent, and a simple retry will get you on your way.

so permit the action to fail the first time,
and then run the action again if that happens.