tests: fix midnight bug
Running the test over midnight
can cause a failure since now()
gets evaluated before midnight,
and the test runs after midnight.
Tuck the expected value in behind
a lambda that is evaluated inside
the function instead so now() is
evaluated closer to when the test
runs.

There is still a window of failure
from the query running before midnight
but the expected running after, but
it's at least smaller.