CI: add security scanner
Explorer and OWS have security
scans of the Docker files, so
add a nightly run of Trivy
in "fs" mode in this repository
as well. The workflow can also
be triggered manually in the actions
tab.

Based on the output in the console,
this workflow appears to scan uv.lock
for vulnerabilities. This workflow
should detect fewer vulnerabilities
than the OWS workflow since that
also catches vulnerabilities in
various Ubuntu packages that are
installed in the image.