Speed up S3 Client tests with mocks
The tests creating an s3_client were taking ~10 seconds to run 3 tests.
They were waiting on a socket connection that was never going to happen,
because they're connecting to a fake region.

This change makes the tests more granular, and adds moto.mock_aws around
the slow client creation.

On my machine it brings the runtime of test_utils_aws from 9.6s to 1.3s.