index: add return types
Add return types to some methods
that were missing them. Switch
the code in the methods to use
unsafe methods since the preceeding
code guarantees that the lookup
will find the name. These changes
save a database lookup when adding
an already existing MetadataType.

The methods in the memory index
already had type annotations, but
change the implementation to use
the unsafe method there too, which
removes a type cast.