Put coverage config in pyproject
Move the coverage config from
a separate file to pyproject.toml.

While moving, add the alembic
version files to be excluded from
coverage, and update the list of
default "ignores" to include
abstract methods and a few other
reasonable defaults. Also switch
to the modern "exclude_also"
directive which does not overwrite
the coverage.py defaults.