PR and release process updates
The background to this, is that I've been frustrated both creating and
merging PRs into ODC Core. This has changed recently for a couple of reasons.
Collectively we're doing a lot better at creating small, tightly focused PRs
that are easy to review and merge. And we're also enjoying a lot more
work from around the world, meaning from different timezones, meaning you
wake up in the morning and have several PRs to review and merge.

The existing process of updating the `whats_new.rst` file in every PR is
a problem because it requires a created PR to get it's number, and then updating
the commits with the PR number placed into a commit to whats_new.

Then doubly bad, is that due to GitHub not supporting fast-forward merges,
every, single, merged, PR, requires, editing, the, next, before it can
be merged.

# So

Instead of requiring the `whats_new.rst` file be updated in every PR,
we record changes using good quality titles on PRs and then have the release
manager update the `whats_new.rst` file when doing a release.

A minor drawback is that we won't have upcoming changes in the documentation
about the next release.

But has the benefit of not requiring a commit/create pr/update pr number/commit/push dance 
for every PR. And not creating merge conflicts in the `whats_new.rst` file
on every PR merge.

See: Issue #1982