Put JSON type aliases in separate file
Sphinx 8.1.3 cannot resolve the cyclic
JsonLike alias, and there are numerous
open issues around these kinds of problems.

Put the JSON type alises in a separate
module, and configure Sphinx to mock
that module. The documentation for
those 3 aliases is lost, but several
big red WARNING: are eliminated when
building the documentation.