Record rasterio/gdal/etc versions from CI runs
These are somewhat available from when the packages were installed,
but when debugging issues it's vital to know what versions of
GDAL/PROJ/rasterio are actually being used, and it's not always
obvious, depending on how rasterio was installed.

The binary wheels from PyPI include GDAL with them, whereas
installing from source with GDAL source available will be different.

Mixing and matching is where you end up in the dragon's lair.