cog: un-nest and align logic
Make the test once and put that
into a variable, and then use
that variable in the places
that depend on the value. This
makes it more obvious that
the definition of path in the first
if-statement corresponds to
the usage elsewhere.

Also un-nest the triple-nested with
and remove an else after return
to reduce indentation levels a bit.