validate_document: add type annotations
The Resource constructor only takes
one argument, so call the helper
method in DRAFT4 instead.

Also hoist the helper function
from the try-except since that
does not raise any exceptions,
and flip the arguments of
the if-statement that creates
the registry and add a type
ignore on the second line.