Replace the automatic spell checker tool.
Previously, we were automatically spell checking this repository using a
pyspelling GitHub action.

The pyspelling action annoyed me in a number of ways. First, there were a
lot of words which it didn't know about and marked as spelling errors,
which meant we had to keep a spelling word list up to date . This is
a lot of manual toil for no particular benefit. Secondly, and the part
that made this extra annoying is that there was no easy way of running
the checker locally. It was really common to create a new pull request
and then have to go back and edit it because GitHub Actions would report
a spelling mistake. Most of our linting tools are set up in pre-commit,
which if you're using it properly runs before your changes are pushed
to GitHub. Having a tight feedback loop is much, much nicer for anyone
contributing to the repository.

codespell is the tool I've chosen as a replacement. It has a
comprehensive dictionary already populated with a lot of technical words
you'd find in a code repository. It also has rules setup to ignore a
lot of things that might not be in a standard dictionary, but which
*look* like code. This means that it's possible to run with very little
configuration and still get useful feedback on typos and the like.

Over the years, a bunch of misspelling have been missed because they were
accidentally added to our wordlist! 🤦🏻