Simplify intersphinx API links
It's not necessary to specifying the source documentation site, Sphinx
is smart enough to find references to fully qualified Python objects.

There's also a neat syntax for shortening the link text to the last
component of the target, by prefixing the target with a tilde '~'.

See https://www.sphinx-doc.org/en/master/usage/referencing.html#syntax
for details.