test_config_tool: enable test for postgis
Switch the test to use the ls8_eo3_product
so the metadata validation performed with
Postgis can pass, and enable the test with
Postgis.

The asserted log messages are never printed
by the Postgis driver, so make the asserts
conditional on the environment and only
test them on the Postgres driver.