drivers: use has_schema method
The inspector has a method for
querying about the existence
of a schema, so use that instead
of scanning the result of getting
all schema names.

Explorer also checks for schemas,
so give the function an argument
for the schema name so Explorer
does not have to re-implement
the same logic.