documents: handle failed download
This function is used by odc-tools,
and the doc_stream can have None
as document when S3 fails, and that
causes json.loads to raise a TypeError
which is not caught by the except
clause.

Fix the type annotation and throw
an error locally when doc is None
so the on_error handler is invoked
properly.